deps: move to Prisma 7 with the pg driver adapter
The riskiest commit in this PR, kept last and self-contained so it can be
dropped with a rebase if you would rather take it separately.

Prisma 7 changes three things that reach this codebase:

1. **Connection URLs leave the schema.** `url` and `directUrl` are no longer
   valid in `datasource db`, so they move to a new `prisma.config.ts`, which
   Migrate reads. The same environment variables are used, in the same order,
   so no deployment needs to change anything: `POSTGRES_PRISMA_URL` and
   `POSTGRES_URL_NON_POOLING` keep their current meaning.

2. **The client needs an explicit driver adapter.** `src/lib/prisma.ts` builds a
   `PrismaPg` adapter over `POSTGRES_PRISMA_URL`. `pg` is already a dependency.

3. **The client is generated into the source tree**, at
   `src/generated/prisma`, rather than into node_modules. 17 imports move from
   `@prisma/client` to `@/generated/prisma/client` (server) or
   `@/generated/prisma/browser` (anything reachable from a client component).
   The directory is gitignored, dockerignored, and excluded from linting.

Two consequences worth calling out, both found by building rather than by
reading the changelog:

- `src/lib/api.ts` exported `randomId`, and two client components imported it
  from there. Under Prisma 6 that was harmless; under Prisma 7 it pulls the pg
  adapter and `pg` itself into the browser bundle and the build fails on Node
  built-ins. `randomId` moves to `src/lib/random.ts` with no Prisma in its
  import graph, and `api.ts` re-exports it so server-side callers are unchanged.
- The browser namespace exports `Prisma.Decimal` as a value only, so
  `src/trpc/client.tsx` derives the instance type with `InstanceType<typeof ...>`
  for its superjson registration.

Dockerfile: `prisma generate` moves after the source copy (the output path is
now inside `src/`), `prisma.config.ts` is copied into both the build and the
runtime stage, and the runtime stage stops regenerating the client — the build
bundles it into `.next`, and that stage has no source tree to generate from.
