+++ .dockerignore
@@ -7,6 +7,9 @@ node_modules
 out
 build
 postgres-data
+# Generated by `prisma generate` inside the image; a local copy would be stale
+# or built for another platform.
+src/generated
 .git
 .github
 .devcontainer
