+++ eslint.config.mjs
@@ -10,6 +10,9 @@ const eslintConfig = defineConfig([
     'out/**',
     'build/**',
     'next-env.d.ts',
+    // Prisma 7 emits the generated client here; it ships its own
+    // eslint-disable banners and should not be linted as project source.
+    'src/generated/**',
   ]),
   {
     rules: {
