+++ package.json
@@ -30,7 +30,8 @@
     "@formatjs/intl-localematcher": "^0.8.13",
     "@hookform/resolvers": "^5",
     "@json2csv/plainjs": "^7.0.8",
-    "@prisma/client": "^6.19.3",
+    "@prisma/adapter-pg": "^7.0.0",
+    "@prisma/client": "^7.0.0",
     "@radix-ui/react-checkbox": "^1.3.11",
     "@radix-ui/react-collapsible": "^1.1.20",
     "@radix-ui/react-dialog": "^1.1.23",
@@ -67,7 +68,7 @@
     "next13-progressbar": "^1.2.2",
     "openai": "^6.0.0",
     "pg": "^8.23.0",
-    "prisma": "^6.19.3",
+    "prisma": "^7.9.1",
     "qrcode.react": "^4.2.0",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
