+++ prisma/schema.prisma
@@ -2,13 +2,15 @@
 // learn more about it in the docs: https://pris.ly/d/prisma-schema
 
 generator client {
-  provider = "prisma-client-js"
+  provider = "prisma-client"
+  output   = "../src/generated/prisma"
 }
 
 datasource db {
-  provider  = "postgresql"
-  url       = env("POSTGRES_PRISMA_URL") // uses connection pooling
-  directUrl = env("POSTGRES_URL_NON_POOLING") // uses a direct connection
+  provider = "postgresql"
+  // Connection URLs moved to prisma.config.ts (Prisma 7 no longer supports
+  // `url`/`directUrl` in the schema). The app passes a driver adapter to
+  // PrismaClient; Migrate reads its connection from prisma.config.ts.
 }
 
 model Group {
