+++ src/app/groups/[groupId]/balances-list.tsx
@@ -1,7 +1,7 @@
+import { Participant } from '@/generated/prisma/browser'
 import { Balances } from '@/lib/balances'
 import { Currency } from '@/lib/currency'
 import { cn, formatCurrency } from '@/lib/utils'
-import { Participant } from '@prisma/client'
 import { useLocale } from 'next-intl'
 
 type Props = {
