+++ src/app/groups/[groupId]/share-button.tsx
@@ -9,8 +9,8 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover'
+import { Group } from '@/generated/prisma/browser'
 import { useBaseUrl } from '@/lib/hooks'
-import { Group } from '@prisma/client'
 import { Share } from 'lucide-react'
 import { useTranslations } from 'next-intl'
