+++ src/components/category-selector.tsx
@@ -16,8 +16,8 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover'
+import { Category } from '@/generated/prisma/browser'
 import { useMediaQuery } from '@/lib/hooks'
-import { Category } from '@prisma/client'
 import { useTranslations } from 'next-intl'
 import { forwardRef, useEffect, useState } from 'react'
