+++ src/lib/balances.ts
@@ -1,6 +1,6 @@
+import { Participant } from '@/generated/prisma/browser'
 import { getGroupExpenses } from '@/lib/api'
 import { getExpenseShares } from '@/lib/shares'
-import { Participant } from '@prisma/client'
 
 export type Balances = Record<
   Participant['id'],
