+++ src/lib/prisma.ts
@@ -1,21 +1,28 @@
-import { PrismaClient } from '@prisma/client'
+import { PrismaClient } from '@/generated/prisma/client'
+import { PrismaPg } from '@prisma/adapter-pg'
 
-// `types: ["node", "jest"]` narrows the ambient globals, so the old
-// `declare const global: Global` is no longer available. `globalThis` is the
-// portable spelling and needs no ambient declaration.
 const globalForPrisma = globalThis as unknown as { prisma?: PrismaClient }
+
+// Prisma 7 ships a Rust-free client and requires an explicit driver adapter.
+// We use the `pg` adapter with the pooled connection string.
+function createPrismaClient() {
+  const adapter = new PrismaPg({
+    connectionString: process.env['POSTGRES_PRISMA_URL'],
+  })
+  return new PrismaClient({
+    adapter,
+    // log: [{ emit: 'stdout', level: 'query' }],
+  })
+}
 
 export let p: PrismaClient = undefined as any as PrismaClient
 
 if (typeof window === 'undefined') {
-  // await delay(1000)
   if (process.env['NODE_ENV'] === 'production') {
-    p = new PrismaClient()
+    p = createPrismaClient()
   } else {
     if (!globalForPrisma.prisma) {
-      globalForPrisma.prisma = new PrismaClient({
-        // log: [{ emit: 'stdout', level: 'query' }],
-      })
+      globalForPrisma.prisma = createPrismaClient()
     }
     p = globalForPrisma.prisma
   }
