+++ src/lib/random.ts
@@ -1,0 +1,5 @@
+import { nanoid } from 'nanoid'
+
+export function randomId() {
+  return nanoid()
+}
