+++ src/lib/shares.ts
@@ -1,4 +1,4 @@
-import { SplitMode } from '@prisma/client'
+import { SplitMode } from '@/generated/prisma/browser'
 import { match } from 'ts-pattern'
 
 /**
